Fix build warning
