+++ app/gui2/src/util/ast/aliasAnalysis.ts
@@ -6,7 +6,7 @@ import {
   parsedTreeOrTokenRange,
   readAstOrTokenSpan,
   readTokenSpan,
-} from '@/util/ast'
+} from '@/util/ast/raw'
 import { MappedKeyMap, MappedSet, NonEmptyStack } from '@/util/containers'
 import type { LazyObject } from 'shared/ast/parserSupport'
 import { rangeIsBefore, sourceRangeKey, type SourceRange } from 'shared/yjsModel'
